+++ module/PSETW.psm1
@@ -3,18 +3,18 @@ $moduleName = [System.IO.Path]::GetFileNameWithoutExtension($PSCommandPath)
 
 if ($IsCoreClr) {
     $isReload = $true
-    if (-not ('PSETW.Shared.LoadContext' -as [type])) {
+    if (-not ('PSEtw.Shared.LoadContext' -as [type])) {
         $isReload = $false
 
         Add-Type -Path ([System.IO.Path]::Combine($PSScriptRoot, 'bin', 'net6.0', "$moduleName.Shared.dll"))
     }
 
-    $mainModule = [PSETW.Shared.LoadContext]::Initialize()
+    $mainModule = [PSEtw.Shared.LoadContext]::Initialize()
     $innerMod = &$importModule -Assembly $mainModule -PassThru:$isReload
 }
 else {
-    $innerMod = if ('PSETW.PSETWGlobals' -as [type]) {
-        $modAssembly = [PSETW.PSETWGlobals].Assembly
+    $innerMod = if ('PSEtw.PSETWGlobals' -as [type]) {
+        $modAssembly = [PSEtw.PSETWGlobals].Assembly
         &$importModule -Assembly $modAssembly -Force -PassThru
     }
     else {
@@ -27,7 +27,7 @@ if ($innerMod) {
     # Bug in pwsh, Import-Module in an assembly will pick up a cached instance
     # and not call the same path to set the nested module's cmdlets to the
     # current module scope. This is only technically needed if someone is
-    # calling 'Import-Module -Name PSETW -Force' a second time. The first
+    # calling 'Import-Module -Name PSEtw -Force' a second time. The first
     # import is still fine.
     # https://github.com/PowerShell/PowerShell/issues/20710
     $addExportedCmdlet = [System.Management.Automation.PSModuleInfo].GetMethod(