+++ tests/units/JordanTest.cs
@@ -1,4 +1,4 @@
-using PSETW;
+using PSEtw;
 using System;
 using Xunit;