+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-# Changelog for PSETW
+# Changelog for PSEtw
 
 ## v0.1.0 - 2023-06-23
 
-+ Initial version of the `PSETW` module
++ Initial version of the `PSEtw` module